Added field positioning.
Fields are now positioned based on their row and col values.

To help with positioning, the cards, field and paths all inherit their
width from the board. That way, we can move them by specifying a 'left'
and 'top' value in multiples of 100%. Resizing the game is a matter of
changing the size of the board.