Tweak to rotation of card.
Ensure the rotation transition takes place over all other cards. To
achieve this, a global zIndex is kept. Whenever a card is clicked on,
it increments this global zIndex and adopts the value for its next
render.

I'm not using state to keep the zIndex of the card, because I do not
want to trigger an immediate rerender for zero gain (the z-index only
matters when the card does a rotation animation, when its corners go over
or under its neighbours).