Changed the way fields and cards are rendered.
Instead of placing card divs directly in field divs, all fields are now
in their own layer, and all cards in a separate layer on top. Because
the cards obtain the same coordinates as the fields they are placed in,
the are placed nicely on top.

The advantage of this is that less DOM rewrites are necessary when cards
are moved. Additionaly, animating cards moving from one field to another
has become much easier (using a simple CSS transition). This opens the
way to move cards by tapping a card, then tapping the field to move to
instead of using drag and drop.