Restructuring of code.
The `view` directory now contains an index file that exports a function
for rendering the world. It also exports the uiEvents stream separately.

The `logic` directory is now called `engine`. It contains an index file
that exports a function for creating a game.

A separate `engine-worker` directory is created. It also contains an
index file with the same signature as the one in `engine`. By requiring
`engine-worker` instead of `engine`, one gets the same engine but
running in a web-worker. This allows easy switching during development.

An explicit difference is made between the `world` (an observable) and
the `worldState` (the world at a certain moment represented by an object).