First working version route calculation.
The algorithm seems to work, but is a bit slow because it operates on
non-optimized data structures. We can now optimize this part.

The route calculation is activated upon hovering for a small amount of
time over a path. This feels fiddly, so a different approach should be
chosen (like highlighting all routes when hovering over a card).