Rewrote engine with an eye on persistence.
The engine has been rewritten to use separate read and worldstate
stores. In memory stores are used, which is fine in the browser but can
easily be replaced with for instance redis on the backend.

TODO: Engine commands need to be made asynchronous.

Moves reference the previous move they follow by hash. World states are
also stored by move hash. This means the engine is capable of running
parallel games by default, also mostly relevant to a backend.

Also, eslint has been updated to a new version.