Add support for in-file unit tests.
Mocha describe and it statements can be used in files and will be
removed by a sweet.js macro during webpack compilation.

Basic functionality is in place, but source maps are broken and for some
reason, macro's cannot be let loose on the root src file without
breaking the app. Further investigation required.